+++ packages/pg-driver/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @slonik/pg-driver
 
+## 46.4.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @slonik/driver@46.4.0
+  - @slonik/errors@46.4.0
+  - @slonik/sql-tag@46.4.0
+  - @slonik/types@46.4.0
+  - @slonik/utilities@46.4.0
+
 ## 46.3.0
 
 ### Patch Changes