+++ packages/slonik-interceptor-query-cache/CHANGELOG.md
@@ -1,5 +1,7 @@
 # slonik-interceptor-query-cache
 
+## 46.4.0
+
 ## 46.3.0
 
 ## 46.2.0